refactor: move create task modal to component folders
